Moved discord tasks to separate cog
