+++ bot.py
@@ -20,12 +20,13 @@ import numpy as np
 from discord import channel
 from discord.ext import commands, tasks
 
-from src.sheets.events import get_events
+from src.discord.utils import auto_report
+# from src.sheets.events import get_events
 # from src.sheets.tournaments import get_tournament_channels
 from src.sheets.censor import get_censor
-from src.sheets.sheets import send_variables, get_variables, get_tags
-from src.forums.forums import open_browser
-from src.wiki.stylist import prettify_templates
+# from src.sheets.sheets import get_tags
+# from src.forums.forums import open_browser
+# from src.wiki.stylist import prettify_templates
 # from src.wiki.tournaments import get_tournament_list
 # from src.wiki.wiki import implement_command, get_page_tables
 # from src.wiki.scilympiad import get_points
@@ -47,8 +48,6 @@ from tournaments import update_tournament_list
 ##############
 # SERVER VARIABLES
 ##############
-
-from src.discord.utils import *
 
 from src.discord.globals import *
 
@@ -92,245 +91,10 @@ bot.remove_command("help")
 async def on_ready():
     """Called when the bot is enabled and ready to be run."""
     print(f'{bot.user} has connected!')
-    try:
-        await pull_prev_info()
-    except Exception as e:
-        print("Error in starting function with pulling previous information:")
-        print(e)
-
-    try:
-        await update_tournament_list(bot)
-    except Exception as e:
-        print("Error in starting function with updating tournament list:")
-        print(e)
-
-    try:
-        refresh_sheet.start()
-    except Exception as e:
-        print("Error in starting function with updating tournament list:")
-        print(e)
-
-    post_something.start()
-    cron.start()
-    go_stylist.start()
-    manage_welcome.start()
-    store_variables.start()
-    change_bot_status.start()
-    update_member_count.start()
     
-@tasks.loop(minutes=5)
-async def update_member_count():
-    """Updates the member count shown on hidden VC"""
-    guild = bot.get_guild(SERVER_ID)
-    channel_prefix = "Members"
-    vc = discord.utils.find(lambda c: channel_prefix in c.name, guild.voice_channels)
-    mem_count = guild.member_count
-    joined_today = len([m for m in guild.members if m.joined_at.date() == datetime.datetime.today().date()])
-    left_channel = discord.utils.get(guild.text_channels, name=CHANNEL_LEAVE)
-    left_messages = await left_channel.history(limit=200).flatten()
-    left_today = len([m for m in left_messages if m.created_at.date() == datetime.datetime.today().date()])
-    await vc.edit(name=f"{mem_count} Members (+{joined_today}/-{left_today})")
-    print("Refreshed member count.")
-
-@tasks.loop(seconds=30.0)
-async def refresh_sheet():
-    """Refreshes the censor list and stores variable backups."""
-    try:
-        await refresh_algorithm()
-    except Exception as e:
-        print("Error when completing the refresh algorithm when refreshing the sheet:")
-        print(e)
-
-    try:
-        await prepare_for_sending()
-    except Exception as e:
-        print("Error when sending variables to log sheet:")
-        print(e)
-
-    print("Attempted to refresh/store data from/to sheet.")
-
-@tasks.loop(hours=10)
-async def store_variables():
-    await prepare_for_sending("store")
-
-@tasks.loop(hours=24)
-async def go_stylist():
-    await prettify_templates()
-
-@tasks.loop(minutes=10)
-async def manage_welcome():
-    server = bot.get_guild(SERVER_ID)
-    now = datetime.datetime.now()
-    # Channel message deleting is currently disabled
-    # if now.hour < ((0 - TZ_OFFSET) % 24) or now.hour > ((11 - TZ_OFFSET) % 24):
-    #     print(f"Cleaning #{CHANNEL_WELCOME}.")
-    #     # if between 12AM EST and 11AM EST do not do the following:
-    #     channel = discord.utils.get(server.text_channels, name=CHANNEL_WELCOME)
-    #     async for message in channel.history(limit=None):
-    #         # if message is over 3 hours old
-    #         author = message.author
-    #         user_no_delete = await is_launcher_no_ctx(message.author)
-    #         num_of_roles = len(author.roles)
-    #         if num_of_roles > 4 and (now - author.joined_at).seconds // 60 > 1 and not user_no_delete:
-    #             await _confirm([author])
-    #         if (now - message.created_at).seconds // 3600 > 3 and not message.pinned:
-    #             # delete it
-    #             await message.delete()
-    # else:
-    #     print(f"Skipping #{CHANNEL_WELCOME} clean because it is outside suitable time ranges.")
-
-@tasks.loop(minutes=1)
-async def cron():
-    print("Executed cron.")
-    global CRON_LIST
-    for c in CRON_LIST:
-        date = c['date']
-        if datetime.datetime.now() > date:
-            # The date has passed, now do
-            CRON_LIST.remove(c)
-            await handle_cron(c['do'])
-
-async def handle_cron(string):
-    try:
-        if string.find("unban") != -1:
-            iden = int(string.split(" ")[1])
-            server = bot.get_guild(SERVER_ID)
-            member = await bot.fetch_user(int(iden))
-            await server.unban(member)
-            print(f"Unbanned user ID: {iden}")
-        elif string.find("unmute") != -1:
-            iden = int(string.split(" ")[1])
-            server = bot.get_guild(SERVER_ID)
-            member = server.get_member(int(iden))
-            role = discord.utils.get(server.roles, name=ROLE_MUTED)
-            self_role = discord.utils.get(server.roles, name=ROLE_SELFMUTE)
-            await member.remove_roles(role, self_role)
-            print(f"Unmuted user ID: {iden}")
-        elif string.find("unstealfishban") != -1:
-            iden = int(string.split(" ")[1])
-            STEALFISH_BAN.remove(iden)
-            print(f"Un-stealfished user ID: {iden}")
-        else:
-            print("ERROR:")
-            await auto_report(bot, "Error with a cron task", "red", f"Error: `{string}`")
-    except Exception as e:
-        await auto_report(bot, "Error with a cron task", "red", f"Error: `{e}`\nOriginal task: `{string}`")
-
-@tasks.loop(hours=1)
-async def change_bot_status():
-    statuses = [
-        {"type": "playing", "message": "Game On"},
-        {"type": "listening", "message": "my SoM instrument"},
-        {"type": "playing", "message": "with Pi-Bot Beta"},
-        {"type": "playing", "message": "with my gravity vehicle"},
-        {"type": "watching", "message": "the WS trials"},
-        {"type": "watching", "message": "birbs"},
-        {"type": "watching", "message": "2018 Nationals again"},
-        {"type": "watching", "message": "the sparkly stars"},
-        {"type": "watching", "message": "over the week"},
-        {"type": "watching", "message": "for tourney results"},
-        {"type": "listening", "message": "birb sounds"},
-        {"type": "playing", "message": "with proteins"},
-        {"type": "playing", "message": "with my detector"},
-        {"type": "playing", "message": "Minecraft"},
-        {"type": "playing", "message": "with circuits"},
-        {"type": "watching", "message": "my PPP fall"},
-        {"type": "playing", "message": "a major scale"},
-        {"type": "listening", "message": "clinking medals"},
-        {"type": "watching", "message": "the world learn"},
-        {"type": "watching", "message": "SciOly grow"},
-        {"type": "watching", "message": "tutorials"},
-        {"type": "playing", "message": "with wiki templates"},
-        {"type": "playing", "message": "the flute"},
-        {"type": "watching", "message": "bear eat users"},
-        {"type": "watching", "message": "xkcd"},
-        {"type": "playing", "message": "with wiki templates"},
-        {"type": "watching", "message": "Jmol tutorials"},
-    ]
-    botStatus = random.choice(statuses)
-    if botStatus["type"] == "playing":
-        await bot.change_presence(activity=discord.Game(name=botStatus["message"]))
-    elif botStatus["type"] == "listening":
-        await bot.change_presence(activity=discord.Activity(type=discord.ActivityType.listening, name=botStatus["message"]))
-    elif botStatus["type"] == "watching":
-        await bot.change_presence(activity=discord.Activity(type=discord.ActivityType.watching, name=botStatus["message"]))
-    print("Changed the bot's status.")
-
-@tasks.loop(hours=28)
-async def post_something():
-    global can_post
-    """Allows Pi-Bot to post markov-generated statements to the forums."""
-    if can_post:
-        print("Attempting to post something.")
-        await open_browser()
-    else:
-        can_post = True
-
-async def refresh_algorithm():
-    """Pulls data from the administrative sheet."""
-    try:
-        global CENSORED_WORDS
-        global CENSORED_EMOJIS
-        censor = await get_censor()
-        CENSORED_WORDS = censor[0]
-        CENSORED_EMOJIS = censor[1]
-    except Exception as e:
-        print("Could not refresh censor in refresh_algorithm:")
-        print(e)
-
-    try:
-        global EVENT_INFO
-        EVENT_INFO = await get_events()
-    except Exception as e:
-        print("Could not refresh event list in refresh_algorithm:")
-        print(e)
-
-    try:
-        global TAGS
-        TAGS = await get_tags()
-    except Exception as e:
-        print("Could not refresh tags in refresh_algorithm:")
-        print(e)
-    
-    print("Refreshed data from sheet.")
-    return True
-
-async def prepare_for_sending(type="variable"):
-    """Sends local variables to the administrative sheet as a backup."""
-    r1 = json.dumps(REPORT_IDS)
-    r2 = json.dumps(PING_INFO)
-    r3 = json.dumps(TOURNEY_REPORT_IDS)
-    r4 = json.dumps(COACH_REPORT_IDS)
-    r5 = json.dumps(CRON_LIST, default = datetime_converter)
-    r6 = json.dumps(REQUESTED_TOURNAMENTS)
-    await send_variables([[r1], [r2], [r3], [r4], [r5], [r6]], type)
-    print("Stored variables in sheet.")
-
 def datetime_converter(o):
     if isinstance(o, datetime.datetime):
         return o.__str__()
-
-async def pull_prev_info():
-    data = await get_variables()
-    global PING_INFO
-    global REPORT_IDS
-    global TOURNEY_REPORT_IDS
-    global COACH_REPORT_IDS
-    global CRON_LIST
-    global REQUESTED_TOURNAMENTS
-    REPORT_IDS = data[0][0]
-    PING_INFO = data[1][0]
-    TOURNEY_REPORT_IDS = data[2][0]
-    COACH_REPORT_IDS = data[3][0]
-    cron = data[4][0]
-    for c in cron:
-        try:
-            c['date'] = datetime.datetime.strptime(c['date'], "%Y-%m-%d %H:%M:%S.%f")
-        except Exception as e:
-            print("ERROR WITH CRON TASK: ", e)
-    CRON_LIST = cron
-    REQUESTED_TOURNAMENTS = data[5][0]
-    print("Fetched previous variables.")
 
 async def get_words():
     """Gets the censor list"""
@@ -861,6 +625,7 @@ bot.load_extension("src.discord.staffcommands")
 bot.load_extension("src.discord.membercommands")
 bot.load_extension("src.discord.devtools")
 bot.load_extension("src.discord.funcommands")
+bot.load_extension("src.discord.tasks")
 
 if dev_mode:
     bot.run(DEV_TOKEN)
